safari-534.53-branch/LayoutTests/platform/chromium-win/fast/inspector-support/uncaught-dom8-exception-expected.txt

CONSOLE MESSAGE: line 9: Uncaught Error: NOT_FOUND_ERR: DOM Exception 8
This tests that we are getting the correct message for DOM Exception 8: NOT_FOUND_ERR.
